fix: remove native repair checkpoint state
Generated-By: looper 0.10.1 (runner=fixer, agent=codex)
